+++ 1.21.6/src/main/java/com/bleudev/nine_lifes/WindChargeTickFeatures.java
@@ -11,18 +11,19 @@ import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.projectile.WindChargeEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.server.world.ServerWorld;
-import net.minecraft.text.Text;
-import net.minecraft.util.TypeFilter;
+import net.minecraft.sound.SoundCategory;
+import net.minecraft.sound.SoundEvents;
 import net.minecraft.util.math.Box;
 import net.minecraft.world.World;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.Objects;
 
 import static com.bleudev.nine_lifes.compat.VersionCompat.getPosCompat;
 
 public class WindChargeTickFeatures {
-    public static void do_for(ServerWorld world, WindChargeEntity wind_charge) {
-        Box action_box = Box.of(getPosCompat(wind_charge), 3, 3, 3);
+    protected static void doWindChargeTick(@NotNull ServerWorld world, WindChargeEntity wind_charge) {
+        Box action_box = NineLifesConst.windChargeActionBox(wind_charge);
         WorldUtils.forBlocksInBox(action_box, (pos) -> {
             var blockEntity = world.getBlockEntity(pos);
             if (blockEntity instanceof BrewingStandBlockEntity brewing) {
@@ -45,8 +46,6 @@ public class WindChargeTickFeatures {
                 );
             }
         });
-
-        System.out.println(TypeFilter.instanceOf(LivingEntity.class));
 
         world.getEntitiesByClass(LivingEntity.class, action_box, ignored -> true).forEach(entity -> {
             entity.removeStatusEffect(CustomEffects.AMETHYSM);
@@ -75,7 +74,7 @@ public class WindChargeTickFeatures {
             }
 
             if (inventory_updated)
-                player.sendMessage(Text.of("Broke potions!"), false);
+                world.playSound(player, player.getX(), player.getY(), player.getZ(), SoundEvents.BLOCK_GLASS_BREAK, SoundCategory.PLAYERS);
         });
     }
 }